int to float corrigé

+++ Assets/Scripts/Ressources.cs
@@ -6,7 +6,7 @@ public class Ressources : MonoBehaviour
 {
 
     public int[] resources = new int[4];
-    private int healthPoints;
+    private float healthPoints;
     
     // Start is called before the first frame update
     void Start()
@@ -16,7 +16,7 @@ public class Ressources : MonoBehaviour
     }
 
 
-    public void GetHit(int damage)
+    public void GetHit(float damage)
     {
         healthPoints -= damage;
         if(healthPoints <= 0)
